#13663: Relocate cb_utils.hpp
