+++ ttnn/cpp/ttnn/operations/data_movement/untilize_with_halo_v2/device/untilize_with_halo_v2_program_factory.cpp
@@ -6,7 +6,7 @@
 
 #include <math.h>
 
-#include "tt_dnn/op_library/cb_utils.hpp"
+#include "ttnn/operations/moreh/cb_utils.hpp"
 #include "ttnn/operations/moreh/math.hpp"
 #include "tt_metal/common/constants.hpp"
 #include "tt_metal/detail/util.hpp"
