+++ ttnn/cpp/ttnn/operations/embedding_backward/device/embedding_backward_program_factory.cpp
@@ -5,7 +5,7 @@
 #include "tt_metal/common/constants.hpp"
 #include "tt_metal/detail/util.hpp"
 #include "tt_metal/host_api.hpp"
-#include "ttnn/deprecated/tt_dnn/op_library/cb_utils.hpp"
+#include "ttnn/operations/moreh/cb_utils.hpp"
 #include "ttnn/operations/moreh/math.hpp"
 #include "tt_metal/common/work_split.hpp"
 #include "ttnn/operations/embedding_backward/device/embedding_backward_device_operation.hpp"
